+++ PGDatePicker/NSDate+PGCategory.m
@@ -9,123 +9,10 @@
 
 @implementation NSDate (PGCategory)
 
-static const NSCalendarUnit unitFlags = NSCalendarUnitYear | NSCalendarUnitMonth | NSCalendarUnitDay | NSCalendarUnitHour | NSCalendarUnitMinute | NSCalendarUnitSecond | NSCalendarUnitWeekday;
-
-+ (NSDate *)setYear:(NSInteger)year {
-    NSCalendar *calendar = [NSCalendar currentCalendar];
-    NSDate *currentDate = [NSDate date];
-    NSDateComponents *components = [calendar components:unitFlags fromDate:currentDate];
-    [components setYear:year];
-    [components setMonth:1];
-    [components setHour:0];
-    [components setMinute:0];
-    [components setSecond:0];
-    NSDate *date = [calendar dateFromComponents:components];
++ (NSDate *)dateFromComponents:(NSDateComponents*)dateComponents {
+    NSCalendar * calendar = [NSCalendar currentCalendar];
+    NSDate * date = [calendar dateFromComponents:dateComponents];
     return date;
-}
-
-+ (NSDate *)setYear:(NSInteger)year month:(NSInteger)month {
-    NSCalendar *calendar = [NSCalendar currentCalendar];
-    NSDate *currentDate = [NSDate date];
-    NSDateComponents *components = [calendar components:unitFlags fromDate:currentDate];
-    [components setYear:year];
-    [components setMonth:month];
-    [components setHour:0];
-    [components setMinute:0];
-    [components setSecond:0];
-    NSDate *date = [calendar dateFromComponents:components];
-    return date;
-}
-
-+ (NSDate *)setYear:(NSInteger)year month:(NSInteger)month day:(NSInteger)day {
-    NSCalendar *calendar = [NSCalendar currentCalendar];
-    NSDate *currentDate = [NSDate date];
-    NSDateComponents *components = [calendar components:unitFlags fromDate:currentDate];
-    [components setYear:year];
-    [components setMonth:month];
-    [components setDay:day];
-    [components setHour:0];
-    [components setMinute:0];
-    [components setSecond:0];
-    NSDate *date = [calendar dateFromComponents:components];
-    return date;
-}
-
-+ (NSDate *)setYear:(NSInteger)year month:(NSInteger)month day:(NSInteger)day hour:(NSInteger)hour minute:(NSInteger)minute {
-    NSCalendar *calendar = [NSCalendar currentCalendar];
-    NSDate *currentDate = [NSDate date];
-    NSDateComponents *components = [calendar components:unitFlags fromDate:currentDate];
-    [components setYear:year];
-    [components setMonth:month];
-    [components setDay:day];
-    [components setHour:hour];
-    [components setMinute:minute];
-    NSDate *date = [calendar dateFromComponents:components];
-    return date;
-}
-
-+ (NSDate *)setYear:(NSInteger)year month:(NSInteger)month day:(NSInteger)day hour:(NSInteger)hour minute:(NSInteger)minute second:(NSInteger)second {
-    NSCalendar *calendar = [NSCalendar currentCalendar];
-    NSDate *currentDate = [NSDate date];
-    NSDateComponents *components = [calendar components:unitFlags fromDate:currentDate];
-    [components setYear:year];
-    [components setMonth:month];
-    [components setDay:day];
-    [components setHour:hour];
-    [components setMinute:minute];
-    [components setSecond:second];
-    NSDate *date = [calendar dateFromComponents:components];
-    return date;
-}
-
-+ (NSDate *)setMonth:(NSInteger)month day:(NSInteger)day hour:(NSInteger)hour minute:(NSInteger)minute {
-    NSCalendar *calendar = [NSCalendar currentCalendar];
-    NSDate *currentDate = [NSDate date];
-    NSDateComponents *components = [calendar components:unitFlags fromDate:currentDate];
-    [components setMonth:month];
-    [components setDay:day];
-    [components setHour:hour];
-    [components setMinute:minute];
-    NSDate *date = [calendar dateFromComponents:components];
-    return date;
-}
-
-+ (NSDate *)setHour:(NSInteger)hour minute:(NSInteger)minute {
-    NSCalendar *calendar = [NSCalendar currentCalendar];
-    NSDate *currentDate = [NSDate date];
-    NSDateComponents *components = [calendar components:unitFlags fromDate:currentDate];
-    [components setHour:hour];
-    [components setMinute:minute];
-    NSDate *date = [calendar dateFromComponents:components];
-    return date;
-}
-
-+ (NSDate *)setMinute:(NSInteger)minute second:(NSInteger)second {
-    NSCalendar *calendar = [NSCalendar currentCalendar];
-    NSDate *currentDate = [NSDate date];
-    NSDateComponents *components = [calendar components:unitFlags fromDate:currentDate];
-    [components setMinute:minute];
-    [components setSecond:second];
-    NSDate *date = [calendar dateFromComponents:components];
-    return date;
-}
-
-+ (NSDate *)setHour:(NSInteger)hour minute:(NSInteger)minute second:(NSInteger)second {
-    NSCalendar *calendar = [NSCalendar currentCalendar];
-    NSDate *currentDate = [NSDate date];
-    NSDateComponents *components = [calendar components:unitFlags fromDate:currentDate];
-    [components setHour:hour];
-    [components setMinute:minute];
-    [components setSecond:second];
-    NSDate *date = [calendar dateFromComponents:components];
-    return date;
-}
-
-- (NSInteger)howManyDaysWithMonth {
-    NSCalendar *calendar = [NSCalendar currentCalendar];
-    NSRange range = [calendar rangeOfUnit:NSCalendarUnitDay inUnit:NSCalendarUnitMonth forDate:self];
-    NSInteger day = range.length;
-    return day;
 }
 
 @end
