+++ PGDatePicker/PGDatePicker+Common.m
@@ -56,17 +56,16 @@
     return nil;
 }
 
-- (NSInteger)howManyDaysWithMonthInThisYear:(NSInteger)year withMonth:(NSInteger)month {
-    if((month == 1) || (month == 3) || (month == 5) || (month == 7) || (month == 8) || (month == 10) || (month == 12))
-        return 31 ;
-    if((month == 4) || (month == 6) || (month == 9) || (month == 11))
-        return 30;
-    if((year % 4 == 1) || (year % 4 == 2) || (year % 4 == 3))
-        return 28;
-    if(year % 400 == 0)
-        return 29;
-    if(year % 100 == 0)
-        return 28;
-    return 29;
+- (NSInteger)daysWithMonthInThisYear:(NSInteger)year withMonth:(NSInteger)month {
+    NSDateFormatter *formatter = [[NSDateFormatter alloc] init];
+    [formatter setDateFormat:@"yyyy-MM"];
+    NSString *dateStr = [NSString stringWithFormat:@"%ld-%ld",year,month];
+    NSDate *date = [formatter dateFromString:dateStr];
+    NSCalendar *calendar = [[NSCalendar alloc]initWithCalendarIdentifier: NSCalendarIdentifierGregorian];
+    NSRange range = [calendar rangeOfUnit:NSCalendarUnitDay
+                                   inUnit: NSCalendarUnitMonth
+                                  forDate:date];
+    return range.length;
+    
 }
 @end
