+++ PGDatePicker/PGDatePicker.h
@@ -8,37 +8,8 @@
 #import <UIKit/UIKit.h>
 #import "NSDate+PGCategory.h"
 #import "UIColor+PGHex.h"
-#import "NSCalendar+PGCurrent.h"
 #import <PGPickerView/PGPickerView.h>
-
-typedef NS_ENUM(NSInteger, PGDatePickerMode) {
-    PGDatePickerModeYear, //年
-    PGDatePickerModeYearAndMonth, //年月
-    PGDatePickerModeDate, //年月日
-    PGDatePickerModeDateHour, //年月日时
-    PGDatePickerModeDateHourMinute, //年月日时分
-    PGDatePickerModeDateHourMinuteSecond, //年月日时分秒
-    PGDatePickerModeMonthDay, //月日
-    PGDatePickerModeMonthDayHour, //月日时
-    PGDatePickerModeMonthDayHourMinute, //月日时分
-    PGDatePickerModeMonthDayHourMinuteSecond, //月日时分秒
-    PGDatePickerModeTime, //时分
-    PGDatePickerModeTimeAndSecond, //时分秒
-    PGDatePickerModeMinuteAndSecond, //分秒
-    PGDatePickerModeDateAndTime, //月日周 时分
-};
-
-typedef NS_ENUM(NSUInteger, PGDatePickerType) {
-    PGDatePickerType1,
-    PGDatePickerType2,
-    PGDatePickerType3,
-};
-
-typedef NS_ENUM(NSUInteger, PGShowUnitType) {
-    PGShowUnitTypeAll,
-    PGShowUnitTypeCenter,
-    PGShowUnitTypeNone,
-};
+#import "PGEnumeration.h"
 
 #define PGDatePickerDeprecated(instead) __attribute__((deprecated(instead)))
